requirements.txt

# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests
requests==2.31.0

# Vertex AI auth (Application Default Credentials)
google-auth==2.27.0
google-cloud-texttospeech==2.17.2